Fix circular package dependency
