+++ packages/core/package.json
@@ -13,7 +13,6 @@
     "prepublishOnly": "yarn build"
   },
   "devDependencies": {
-    "@delirvfx/core-test-helper": "*",
     "@types/audiobuffer-to-wav": "1.0.0",
     "@types/jest": "24.0.25",
     "@types/joi": "14.3.4",
