+++ packages/core/src/Engine/Task/EffectRenderTask.spec.ts
@@ -1,8 +1,8 @@
-import { mockClip, mockComposition, mockEffect, mockLayer, mockProject } from '@delirvfx/core-test-helper'
 import { Clip, Project } from '../../Entity'
 import { EffectPluginMissingException } from '../../Exceptions'
 import PluginRegistry from '../../PluginSupport/PluginRegistry'
 import EffectPluginBase from '../../PluginSupport/PostEffectBase'
+import { mockClip, mockComposition, mockEffect, mockLayer, mockProject } from '../../SpecHelper'
 import DependencyResolver from '../DependencyResolver'
 import { RenderContextBase } from '../RenderContext/RenderContextBase'
 import EffectRenderTask from './EffectRenderTask'
