+++ packages/core/src/Exporter.spec.ts
@@ -1,13 +1,5 @@
-import {
-  mockAsset,
-  mockClip,
-  mockComposition,
-  mockEffect,
-  mockKeyframe,
-  mockLayer,
-  mockProject,
-} from '@delirvfx/core-test-helper'
 import { Asset, Clip, Composition, Effect, Keyframe, Layer, Project } from './Entity'
+import { mockAsset, mockClip, mockComposition, mockEffect, mockKeyframe, mockLayer, mockProject } from './SpecHelper'
 import AssetPointer from './Values/AssetPointer'
 import ColorRGB from './Values/ColorRGB'
 import Expression from './Values/Expression'
